requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
numpy==1.16.2
Pillow==6.0.0
pyTelegramBotAPI==3.6.6
requests==2.21.0
six==1.12.0
urllib3==1.24.1
